# Optimize Stack<T>'s enumerator