--- a/src/libraries/System.Collections/tests/Generic/Stack/Stack.Generic.Tests.cs
+++ b/src/libraries/System.Collections/tests/Generic/Stack/Stack.Generic.Tests.cs
@@ -53,7 +53,7 @@ namespace System.Collections.Tests
         protected override bool Contains(IEnumerable<T> enumerable, T value) => ((Stack<T>)enumerable).Contains(value);
         protected override void CopyTo(IEnumerable<T> enumerable, T[] array, int index) => ((Stack<T>)enumerable).CopyTo(array, index);
         protected override bool Remove(IEnumerable<T> enumerable) { ((Stack<T>)enumerable).Pop(); return true; }
-        protected override bool Enumerator_Current_UndefinedOperation_Throws => true;
+        protected override bool Enumerator_Empty_Current_UndefinedOperation_Throws => true;
 
         #endregion