--- a/src/libraries/System.Collections/tests/Generic/Stack/Stack.Tests.cs
+++ b/src/libraries/System.Collections/tests/Generic/Stack/Stack.Tests.cs
@@ -24,7 +24,7 @@ namespace System.Collections.Tests
         }
 
         protected override bool Enumerator_Empty_UsesSingletonInstance => true;
-        protected override bool Enumerator_Current_UndefinedOperation_Throws => true;
+        protected override bool Enumerator_Empty_Current_UndefinedOperation_Throw => true;
         protected override bool Enumerator_Empty_ModifiedDuringEnumeration_ThrowsInvalidOperationException => false;
 
         protected override Type ICollection_NonGeneric_CopyTo_IndexLargerThanArrayCount_ThrowType => typeof(ArgumentOutOfRangeException);