There's a non-trivial amount of unnecessary overhead/complexity in `Stack<T>`'s enumerator. This streamlines it.

(It does also change the undefined behavior of using Current erroneously. We can put back most of its current semantics if we really care. This change puts it on a plan closer to `List<T>`.)

| Method            | Toolchain         | Mean      | Error    | StdDev   | Ratio | Allocated | Alloc Ratio |
|------------------ |------------------ |----------:|---------:|---------:|------:|----------:|------------:|
| IntsEnumerable    | \main\corerun.exe | 232.50 ns | 1.022 ns | 0.956 ns |  1.00 |      40 B |        1.00 |
| IntsEnumerable    | \pr\corerun.exe   |  57.13 ns | 0.401 ns | 0.356 ns |  0.25 |         - |        0.00 |
|                   |                   |           |          |          |       |           |             |
| IntsStack         | \main\corerun.exe | 220.76 ns | 1.008 ns | 0.943 ns |  1.00 |         - |          NA |
| IntsStack         | \pr\corerun.exe   |  54.96 ns | 0.377 ns | 0.335 ns |  0.25 |         - |          NA |
|                   |                   |           |          |          |       |           |             |
| StringsEnumerable | \main\corerun.exe | 345.99 ns | 1.714 ns | 1.604 ns |  1.00 |      40 B |        1.00 |
| StringsEnumerable | \pr\corerun.exe   |  67.93 ns | 0.639 ns | 0.598 ns |  0.20 |         - |        0.00 |
|                   |                   |           |          |          |       |           |             |
| StringsStack      | \main\corerun.exe | 243.20 ns | 1.808 ns | 1.603 ns |  1.00 |         - |          NA |
| StringsStack      | \pr\corerun.exe   |  55.86 ns | 0.603 ns | 0.534 ns |  0.23 |         - |          NA |

```C#
using BenchmarkDotNet.Attributes;
using BenchmarkDotNet.Running;

BenchmarkSwitcher.FromAssembly(typeof(Bench).Assembly).Run(args);

[MemoryDiagnoser(false)]
public class Bench
{
    const int Length = 100;

    private IEnumerable<int> _intsEnumerable = new Stack<int>(Enumerable.Range(0, Length));
    private Stack<int> _intsStack = new Stack<int>(Enumerable.Range(0, Length));
    private IEnumerable<string> _stringsEnumerable = new Stack<string>(Enumerable.Range(0, Length).Select(i => i.ToString()));
    private Stack<string> _stringsStack = new Stack<string>(Enumerable.Range(0, Length).Select(i => i.ToString()));

    [Benchmark]
    public int IntsEnumerable()
    {
        int sum = 0;
        foreach (var i in _intsEnumerable) sum += i;
        return sum;
    }

    [Benchmark]
    public int IntsStack()
    {
        int sum = 0;
        foreach (var i in _intsStack) sum += i;
        return sum;
    }

    [Benchmark]
    public int StringsEnumerable()
    {
        int sum = 0;
        foreach (var s in _stringsEnumerable) sum += s.Length;
        return sum;
    }

    [Benchmark]
    public int StringsStack()
    {
        int sum = 0;
        foreach (var s in _stringsStack) sum += s.Length;
        return sum;
    }
}
```